Feedback Review: moved common queries from tests.__init__.py  to sample_search_queries.py, added discriptive to test_search_returns_error_when_invalid_query_used.
